refactor: #98 - Divide tabela existente em vários arquivos
