+++ client/src/app/(main)/students/loading.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { DataTableSkeleton } from '@/features/students/components/table-skeleton';
+import { DataTableSkeleton } from '@/features/student-list/components/table-skeleton';
 
 export default function Loading() {
   return <DataTableSkeleton columnCount={4} filterableColumnCount={2} />;
